Move User to model package
