+++ gnd/src/main/java/com/google/android/gnd/persistence/local/room/UserEntity.java
@@ -20,7 +20,7 @@ import androidx.annotation.NonNull;
 import androidx.room.ColumnInfo;
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
-import com.google.android.gnd.system.AuthenticationManager.User;
+import com.google.android.gnd.model.User;
 import com.google.auto.value.AutoValue;
 import com.google.auto.value.AutoValue.CopyAnnotations;
