+++ gnd/src/main/java/com/google/android/gnd/repository/ProjectRepository.java
@@ -19,10 +19,10 @@ package com.google.android.gnd.repository;
 import android.util.Log;
 import androidx.annotation.NonNull;
 import com.google.android.gnd.model.Project;
+import com.google.android.gnd.model.User;
 import com.google.android.gnd.persistence.local.LocalDataStore;
 import com.google.android.gnd.persistence.local.LocalValueStore;
 import com.google.android.gnd.persistence.remote.RemoteDataStore;
-import com.google.android.gnd.system.AuthenticationManager.User;
 import io.reactivex.Flowable;
 import io.reactivex.Single;
 import io.reactivex.processors.BehaviorProcessor;
