+++ gnd/src/main/java/com/google/android/gnd/ui/projectselector/ProjectSelectorViewModel.java
@@ -19,6 +19,7 @@ package com.google.android.gnd.ui.projectselector;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.LiveDataReactiveStreams;
 import com.google.android.gnd.model.Project;
+import com.google.android.gnd.model.User;
 import com.google.android.gnd.repository.Loadable;
 import com.google.android.gnd.repository.ProjectRepository;
 import com.google.android.gnd.system.AuthenticationManager;
@@ -36,8 +37,7 @@ public class ProjectSelectorViewModel extends AbstractViewModel {
   ProjectSelectorViewModel(ProjectRepository projectRepository, AuthenticationManager authManager) {
     this.projectRepository = projectRepository;
 
-    AuthenticationManager.User user =
-        authManager.getUser().blockingFirst(AuthenticationManager.User.ANONYMOUS);
+    User user = authManager.getUser().blockingFirst(User.ANONYMOUS);
 
     this.projectSummaries =
         LiveDataReactiveStreams.fromPublisher(projectRepository.getProjectSummaries(user));
